split proxy.go
