+++ server/proxy/call_test.go
@@ -1,0 +1,43 @@
+package proxy
+
+import (
+	"testing"
+
+	"github.com/golang/mock/gomock"
+	"github.com/stretchr/testify/require"
+
+	"github.com/mattermost/mattermost-plugin-apps/apps"
+)
+
+func TestAppMetadataForClient(t *testing.T) {
+	testApps := []apps.App{
+		{
+			BotUserID:   "botid",
+			BotUsername: "botusername",
+			DeployType:  apps.DeployBuiltin,
+			Manifest: apps.Manifest{
+				AppID:       apps.AppID("app1"),
+				DisplayName: "App 1",
+			},
+		},
+	}
+
+	ctrl := gomock.NewController(t)
+	p := newTestProxy(t, testApps, ctrl)
+	creq := apps.CallRequest{
+		Context: apps.Context{
+			UserAgentContext: apps.UserAgentContext{
+				AppID: "app1",
+			},
+		},
+		Call: apps.Call{
+			Path: "/",
+		},
+	}
+
+	resp := p.Call(Incoming{}, creq)
+	require.Equal(t, resp.AppMetadata, AppMetadataForClient{
+		BotUserID:   "botid",
+		BotUsername: "botusername",
+	})
+}
