+++ server/proxy/service_test.go
@@ -19,39 +19,6 @@ import (
 	"github.com/mattermost/mattermost-plugin-apps/upstream"
 )
 
-func TestAppMetadataForClient(t *testing.T) {
-	testApps := []apps.App{
-		{
-			BotUserID:   "botid",
-			BotUsername: "botusername",
-			DeployType:  apps.DeployBuiltin,
-			Manifest: apps.Manifest{
-				AppID:       apps.AppID("app1"),
-				DisplayName: "App 1",
-			},
-		},
-	}
-
-	ctrl := gomock.NewController(t)
-	p := newTestProxy(t, testApps, ctrl)
-	creq := apps.CallRequest{
-		Context: apps.Context{
-			UserAgentContext: apps.UserAgentContext{
-				AppID: "app1",
-			},
-		},
-		Call: apps.Call{
-			Path: "/",
-		},
-	}
-
-	resp := p.Call(Incoming{}, creq)
-	require.Equal(t, resp.AppMetadata, AppMetadataForClient{
-		BotUserID:   "botid",
-		BotUsername: "botusername",
-	})
-}
-
 func newTestProxy(tb testing.TB, testApps []apps.App, ctrl *gomock.Controller) *Proxy {
 	conf := config.NewTestConfigService(nil).WithMattermostConfig(model.Config{
 		ServiceSettings: model.ServiceSettings{
